// frontend/radvis/src/app/viewer/anpassungswunsch/models/anpassungswunsch.ts
/*
 * Copyright (c) 2023 WPS - Workplace Solutions GmbH
 *
 * Licensed under the EUPL, Version 1.2 or as soon they will be approved by the European Commission - subsequent versions of the EUPL (the "Licence");
 *
 * You may not use this work except in compliance with the Licence.
 * You may obtain a copy of the Licence at:
 *
 * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the Licence is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the Licence for the specific language governing permissions and limitations under the Licence.
 */

import { BenutzerName } from 'src/app/shared/models/benutzer-name';
import { PointGeojson } from 'src/app/shared/models/geojson-geometrie';
import { Verwaltungseinheit } from 'src/app/shared/models/verwaltungseinheit';
import { AnpassungswunschKategorie } from 'src/app/viewer/anpassungswunsch/models/anpassungswunsch-kategorie';
import { FehlerprotokollView } from 'src/app/fehlerprotokoll/models/fehlerprotokoll-view';

export interface Anpassungswunsch {
  id: number;
  geometrie: PointGeojson;
  beschreibung: string;
  status: string;
  erstellung: string;
  aenderung: string;
  benutzerLetzteAenderung: BenutzerName;
  verantwortlicheOrganisation: Verwaltungseinheit | null;
  kategorie: AnpassungswunschKategorie;
  ursaechlicheKonsistenzregelVerletzung?: FehlerprotokollView;
  basiertAufKonsistenzregelVerletzung: boolean;
  canEdit: boolean;
}

// eslint-disable-next-line @typescript-eslint/no-namespace
export namespace Anpassungswunsch {
  export const BESCHREIBUNG_MAX_LENGTH = 1000;
}